finish executor events delegate
