+++ azkaban-common/src/main/java/azkaban/executor/JdbcExecutorLoader.java
@@ -34,7 +34,6 @@ import java.sql.SQLException;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -52,6 +51,7 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   private final ExecutorDao executorDao;
   private final ExecutionJobDao executionJobDao;
   private final ExecutionLogsDao executionLogsDao;
+  private final ExecutorEventsDao executorEventsDao;
   private EncodingType defaultEncodingType = EncodingType.GZIP;
 
   @Inject
@@ -59,12 +59,14 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
                             final ExecutionFlowDao executionFlowDao,
                             final ExecutorDao executorDao,
                             final ExecutionJobDao executionJobDao,
-                            final ExecutionLogsDao executionLogsDao) {
+                            final ExecutionLogsDao executionLogsDao,
+                            final ExecutorEventsDao executorEventsDao) {
     super(props, commonMetrics);
     this.executionFlowDao = executionFlowDao;
     this.executorDao = executorDao;
     this.executionJobDao = executionJobDao;
     this.executionLogsDao= executionLogsDao;
+    this.executorEventsDao = executorEventsDao;
   }
 
   public EncodingType getDefaultEncodingType() {
@@ -490,17 +492,8 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   @Override
   public void postExecutorEvent(final Executor executor, final EventType type, final String user,
                                 final String message) throws ExecutorManagerException{
-    final QueryRunner runner = createQueryRunner();
 
-    final String INSERT_PROJECT_EVENTS =
-      "INSERT INTO executor_events (executor_id, event_type, event_time, username, message) values (?,?,?,?,?)";
-    final Date updateDate = new Date();
-    try {
-      runner.update(INSERT_PROJECT_EVENTS, executor.getId(), type.getNumVal(),
-        updateDate, user, message);
-    } catch (final SQLException e) {
-      throw new ExecutorManagerException("Failed to post executor event", e);
-    }
+    this.executorEventsDao.postExecutorEvent(executor, type, user, message);
   }
 
   /**
@@ -512,20 +505,8 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
   @Override
   public List<ExecutorLogEvent> getExecutorEvents(final Executor executor, final int num,
                                                   final int offset) throws ExecutorManagerException {
-    final QueryRunner runner = createQueryRunner();
 
-    final ExecutorLogsResultHandler logHandler = new ExecutorLogsResultHandler();
-    List<ExecutorLogEvent> events = null;
-    try {
-      events =
-        runner.query(ExecutorLogsResultHandler.SELECT_EXECUTOR_EVENTS_ORDER,
-          logHandler, executor.getId(), num, offset);
-    } catch (final SQLException e) {
-      throw new ExecutorManagerException(
-        "Failed to fetch events for executor id : " + executor.getId(), e);
-    }
-
-    return events;
+    return this.executorEventsDao.getExecutorEvents(executor, num, offset);
   }
 
   /**
@@ -861,36 +842,4 @@ public class JdbcExecutorLoader extends AbstractJdbcLoader implements
     }
   }
 
- /**
-   * JDBC ResultSetHandler to fetch records from executor_events table
-   */
-  private static class ExecutorLogsResultHandler implements
-    ResultSetHandler<List<ExecutorLogEvent>> {
-    private static final String SELECT_EXECUTOR_EVENTS_ORDER =
-      "SELECT executor_id, event_type, event_time, username, message FROM executor_events "
-        + " WHERE executor_id=? ORDER BY event_time LIMIT ? OFFSET ?";
-
-    @Override
-    public List<ExecutorLogEvent> handle(final ResultSet rs) throws SQLException {
-      if (!rs.next()) {
-        return Collections.<ExecutorLogEvent> emptyList();
-      }
-
-      final ArrayList<ExecutorLogEvent> events = new ArrayList<>();
-      do {
-        final int executorId = rs.getInt(1);
-        final int eventType = rs.getInt(2);
-        final Date eventTime = rs.getDate(3);
-        final String username = rs.getString(4);
-        final String message = rs.getString(5);
-
-        final ExecutorLogEvent event =
-          new ExecutorLogEvent(executorId, username, eventTime,
-            EventType.fromInteger(eventType), message);
-        events.add(event);
-      } while (rs.next());
-
-      return events;
-    }
-  }
 }
