+++ azkaban-common/src/test/java/azkaban/executor/JdbcExecutorLoaderTest.java
@@ -1061,7 +1061,7 @@ public class JdbcExecutorLoaderTest {
     //TODO kunkun-tang: temporary work-around here. This Test is to be deprecated.
     return new JdbcExecutorLoader(props,
         new CommonMetrics(new MetricsManager(new MetricRegistry())), null
-        , null, null, null);
+        , null, null, null, null);
   }
 
   private boolean isTestSetup() {
